# openapi.yaml
openapi: 3.0.3
info:
  title: Sample Python API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /:
    get:
      summary: Read Root
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  Hello:
                    type: string
                    example: World
  /orders/:
    post:
      summary: Create Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    get:
      summary: Get Orders
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders/{order_id}:
    get:
      summary: Get Order
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
    put:
      summary: Update Order
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
    delete:
      summary: Delete Order
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Order deleted
        '404':
          description: Order not found
components:
  schemas:
    Order:
      type: object
      required:
        - id
        - customer_name
        - items
        - status
      properties:
        id:
          type: string
          format: uuid
        customer_name:
          type: string
        items:
          type: array
          items:
            type: string
        status:
          type: string